Renamed controller and moved IAction results
